payments without pay and invoice details in student side

+++ src/app/pages/dashboard/shared/pages/course-details/shared/pages/course-assignments/course-assignments.component.ts
@@ -1,21 +1,33 @@
 import { Component, inject, OnDestroy } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import {DatePipe, Location, NgForOf, NgIf} from '@angular/common';
-import { LangChangeEvent, TranslatePipe } from '@ngx-translate/core';
+import { DatePipe, Location, NgForOf, NgIf } from '@angular/common';
+import {
+  LangChangeEvent,
+  TranslatePipe,
+  TranslateService,
+} from '@ngx-translate/core';
 import { CourseGroupsService } from '../course-groups/course-groups.service';
-import {ActivatedRoute, Router, RouterLink} from '@angular/router';
+import { ActivatedRoute, Router, RouterLink } from '@angular/router';
 import { ConfirmationService } from 'primeng/api';
 import { CourseRequestModel } from '../../../../../models/course-request.model';
 import { GroupsResponseModel } from '../../../../../models/groups-response.model';
 import { CourseAssignmentsService } from './course-assignments.service';
 import { AssignmentsRequestModel } from './shared/models/assignments-request.model';
-import {Popover} from 'primeng/popover';
-import {TopicRequestModel} from '../../../../../models/topic-request.model';
-import {AssignmentsResponseModel} from './shared/models/assignments-response.model';
+import { Popover } from 'primeng/popover';
+import { TopicRequestModel } from '../../../../../models/topic-request.model';
+import { AssignmentsResponseModel } from './shared/models/assignments-response.model';
 
 @Component({
   selector: 'app-course-assignments',
-  imports: [FormsModule, NgForOf, NgIf, TranslatePipe, Popover, DatePipe, RouterLink],
+  imports: [
+    FormsModule,
+    NgForOf,
+    NgIf,
+    TranslatePipe,
+    Popover,
+    DatePipe,
+    RouterLink,
+  ],
   templateUrl: './course-assignments.component.html',
   styleUrl: './course-assignments.component.scss',
 })
@@ -26,6 +38,7 @@ export class CourseAssignmentsComponent implements OnDestroy {
   private router: Router = inject(Router);
   private confirmationService: ConfirmationService =
     inject(ConfirmationService);
+  private translate: TranslateService = inject(TranslateService);
 
   id = this.route.parent?.snapshot.paramMap.get('id') as string;
 
@@ -54,7 +67,7 @@ export class CourseAssignmentsComponent implements OnDestroy {
 
   list: AssignmentsResponseModel[] = [];
   filteredList: AssignmentsResponseModel[] = [];
-  assignment:AssignmentsResponseModel = new AssignmentsResponseModel();
+  assignment: AssignmentsResponseModel = new AssignmentsResponseModel();
   searchByName() {
     this.filteredList = this.list.filter((obj) =>
       obj.title.toLowerCase().includes(this.searchText.toLowerCase()),
@@ -75,14 +88,35 @@ export class CourseAssignmentsComponent implements OnDestroy {
   }
 
   editItem() {
-    this.service.getItem()
+    this.service.getItem();
   }
 
-  deleteItem() {
-    this.service.deleteItem()
+  delete() {
+    this.confirm('Are you sure you want to delete this assignment?', () => {
+      this.service.deleteItem();
+    });
   }
 
-  getDetails() {
-    this.router.navigate(['/main/educator/dashboard']);
+  confirm(message: string, success: any) {
+    this.confirmationService.confirm({
+      header: this.translate.instant('Confirmation'),
+      message: message,
+      icon: 'pi pi-exclamation-circle',
+      rejectButtonProps: {
+        label: this.translate.instant('Cancel'),
+        icon: 'pi pi-times',
+        outlined: true,
+        size: 'small',
+      },
+      acceptButtonProps: {
+        label: this.translate.instant('Confirm'),
+        icon: 'pi pi-check',
+        size: 'small',
+      },
+      accept: () => {
+        success();
+      },
+      reject: () => {},
+    });
   }
 }
